add Visibility.kt, Id type change from Int to Long

+++ app/src/main/java/work/kcs_labo/idea_telescope/domain/model/user/UserTheme.kt
@@ -1,6 +1,6 @@
 package work.kcs_labo.idea_telescope.domain.model.user
 
 data class UserTheme(
-  val themeId: Int = 0
+  val themeId: Long = 0
 ) {
 }
